Removed legacy window decorator, fixed issues with native decorator fullscreen
