+++ octgnFX/Octgn/Controls/WindowDecorators/NativeDecorator.cs
@@ -1,4 +1,8 @@
-﻿namespace Octgn.Controls.WindowDecorators
+﻿using System;
+using System.ComponentModel;
+using System.Windows;
+
+namespace Octgn.Controls.WindowDecorators
 {
     class NativeDecorator : WindowDecorator
     {
@@ -9,11 +13,34 @@
 
         public override void Apply()
         {
+            Decorated.PropertyChanged += DecoratedOnPropertyChanged;
         }
 
         public override bool Undo()
         {
+            Decorated.PropertyChanged -= DecoratedOnPropertyChanged;
             return true;
         }
+
+        private void DecoratedOnPropertyChanged(object sender, PropertyChangedEventArgs args)
+        {
+            switch (args.PropertyName)
+            {
+
+                case "TitleBarVisibility":
+                    switch (Decorated.TitleBarVisibility)
+                    {
+                        case Visibility.Visible:
+                            Decorated.WindowStyle = WindowStyle.SingleBorderWindow;
+                            break;
+                        case Visibility.Hidden:
+                        case Visibility.Collapsed:
+                            Decorated.WindowStyle = WindowStyle.None;
+                            break;
+                    }
+                    break;
+            }
+        }
+
     }
 }
